feat: jwt_required decorators for endpoints

+++ app/record/record_routes.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, request, jsonify
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.record.record_service import create_record_service, get_record_by_id_service, delete_record_service, \
     get_filtered_records_service
 from app.record.record_schema import RecordSchema
@@ -9,8 +10,12 @@ records_schema = RecordSchema(many=True)
 
 
 @record_bp.route('/record', methods=['POST'])
+@jwt_required()
 def create_record():
     data = request.get_json()
+    current_user_id = get_jwt_identity()
+
+    data['user_id'] = current_user_id
 
     errors = record_schema.validate(data)
     if errors:
@@ -24,24 +29,40 @@ def create_record():
 
 
 @record_bp.route('/record/<record_id>', methods=['GET'])
+@jwt_required()
 def get_record(record_id):
     record = get_record_by_id_service(record_id)
     if not record:
         return jsonify({"error": "Record not found"}), 404
+
+    current_user_id = get_jwt_identity()
+    if record.user_id != current_user_id:
+        return jsonify({"error": "Access denied"}), 403
+
     return jsonify(record_schema.dump(record))
 
 
 @record_bp.route('/record/<record_id>', methods=['DELETE'])
+@jwt_required()
 def delete_record(record_id):
-    record = delete_record_service(record_id)
+    current_user_id = get_jwt_identity()
+
+    record = get_record_by_id_service(record_id)
     if not record:
         return jsonify({"error": "Record not found"}), 404
-    return jsonify(record_schema.dump(record))
+
+    if record.user_id != current_user_id:
+        return jsonify({"error": "Access denied"}), 403
+
+    deleted_record = delete_record_service(record_id)
+    return jsonify(record_schema.dump(deleted_record))
 
 
 @record_bp.route('/record', methods=['GET'])
+@jwt_required()
 def get_records():
-    user_id = request.args.get('user_id')
+    current_user_id = get_jwt_identity()
     category_id = request.args.get('category_id')
-    records = get_filtered_records_service(user_id, category_id)
+
+    records = get_filtered_records_service(current_user_id, category_id)
     return jsonify(records_schema.dump(records))
